finished the rerun

+++ include/User.h
@@ -30,7 +30,7 @@ class User{
     virtual User* clone()=0;
     void addToHistory(Watchable* watch);
     void setName(std::string & name);
-    void setLastRec(int i);
+    void setLastRec(long i);
     std::string& getName();
     bool isInHistory(Watchable& watch) const;
